+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 ### Changed
 
 - nostr: rework `NostrParser` ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/899)
+- nostr: enhance `NostrParser` with flexible parsing options ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/912)
 - pool: refine notification sending depending on event database saving status ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/911)
 
 ### Added